Make Object, Verb, Subject independent components
