+++ react/components/Conditions/Statement.js
@@ -1,9 +1,11 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Button from '../Button'
-import Dropdown from '../Dropdown'
-import Input from '../Input'
 import IconClose from '../icon/Close'
+
+import SubjectAtom from './Atoms/SubjectAtom'
+import VerbAtom from './Atoms/VerbAtom'
+import ObjectAtom from './Atoms/ObjectAtom'
 
 class Statement extends React.Component {
   constructor(props) {
@@ -13,40 +15,6 @@ class Statement extends React.Component {
       error: null,
     }
   }
-
-  static EmptyObject = () => (
-    <div className="flex-auto">
-      <div className="mh3 mb3">
-        <Input key={'object-0'} disabled />
-      </div>
-    </div>
-  )
-
-  static Subject = props => (
-    <div className="flex-auto">
-      <div className={`mh3 ${props.isFullWidth ? 'pb3' : ''}`}>
-        <Dropdown
-          options={props.options}
-          value={!props.condition.subject ? '' : props.condition.subject || ''}
-          onChange={(e, value) => props.onChange(value)}
-        />
-      </div>
-    </div>
-  )
-
-  static Verb = props => (
-    <div className={`mh3 ${props.isFullWidth ? 'pb3' : ''}`}>
-      <Dropdown
-        disabled={!props.condition.subject}
-        options={props.verbs}
-        value={!props.condition.subject ? '' : props.condition.verb || ''}
-        onChange={(e, value) => {
-          const foundVerb = props.verbs.find(verb => verb.value === value)
-          props.onChange(foundVerb)
-        }}
-      />
-    </div>
-  )
 
   handleChangeStatement = (newValue, structure) => {
     this.props.onChangeStatement(newValue, structure)
@@ -81,147 +49,47 @@ class Statement extends React.Component {
     this.handleChangeStatement(null, 'error')
   }
 
-  renderSubject = entities => {
-    const { choices, statements, isFullWidth, statementIndex } = this.props
-    const condition = statements[statementIndex]
-
-    const options = Object.keys(choices).map(choiceKey => {
-      return {
-        value: choiceKey,
-        label: choices[choiceKey].label,
-        unique: choices[choiceKey].unique || false,
-      }
-    })
-
-    const subjectsInUse = statements
-      .map(statement => {
-        return statement.subject
-      })
-      .filter(subject => {
-        return subject !== ''
-      })
-
-    const uniqueOptions = options.filter(option => {
-      if (!option.unique) {
-        return true
-      }
-
-      const alreadyInUse = subjectsInUse.indexOf(option.value) > -1
-      if (!alreadyInUse) {
-        return true
-      }
-
-      if (subjectsInUse.indexOf(option.value) === statementIndex) {
-        return true
-      }
-
-      return false
-    })
-
-    entities.push(
-      <Statement.Subject
-        condition={condition}
-        options={uniqueOptions}
-        isFullWidth={isFullWidth}
-        onChange={selectedSubjectValue => {
-          this.handleChangeStatement(selectedSubjectValue, 'subject')
-          this.clearPredicate()
-        }}
-      />
-    )
-
-    return entities
-  }
-
-  renderVerbs = entities => {
-    const { statements, choices, isFullWidth, statementIndex } = this.props
-    const condition = statements[statementIndex]
-    const myChoice = this.getChoiceBySubject(condition.subject)
-    const desiredVerbs = []
-
-    desiredVerbs.push(
-      <Statement.Verb
-        condition={condition}
-        choices={choices}
-        isFullWidth={isFullWidth}
-        verbs={(myChoice && myChoice.verbs) || [{ label: '', value: '' }]}
-        onChange={verb => {
-          this.handleChangeStatement(verb.value, 'verb')
-          this.clearObjects()
-        }}
-      />
-    )
-
-    entities.push(<div className="flex-auto">{desiredVerbs}</div>)
-
-    return entities
-  }
-
-  renderObjects = entities => {
-    const { statements, statementIndex, isFullWidth } = this.props
-    const condition = statements[statementIndex]
-    const myChoice = this.getChoiceBySubject(condition.subject)
-
-    if (!condition.verb) {
-      entities.push(<Statement.EmptyObject />)
-      return entities
-    }
-
-    if (!myChoice) {
-      entities.push(<Statement.EmptyObject />)
-      return entities
-    }
-
-    const currentVerb = myChoice.verbs.find(
-      verb => verb.value === condition.verb
-    )
-
-    if (!currentVerb) {
-      entities.push(<Statement.EmptyObject />)
-      return entities
-    }
-
-    entities.push(
-      <div className="mh3 flex-auto">
-        {currentVerb.object({
-          statementIndex: statementIndex,
-          statements: statements,
-          isFullWidth: isFullWidth,
-          values: condition.object,
-          error: null,
-        })}
-      </div>
-    )
-
-    return entities
-  }
-
   render() {
     const {
       canDelete,
-      statements,
+      choices,
       isRtl,
       isFullWidth,
+      statements,
       statementIndex,
       labels,
     } = this.props
     const condition = statements[statementIndex]
-    const order = isRtl ? 'OVS' : 'SVO'
-    let statementAtoms = []
-
-    order.split('').map(entity => {
-      if (entity === 'S') {
-        statementAtoms = this.renderSubject(statementAtoms)
-      }
-
-      if (entity === 'V') {
-        statementAtoms = this.renderVerbs(statementAtoms)
-      }
-
-      if (entity === 'O') {
-        statementAtoms = this.renderObjects(statementAtoms)
-      }
-    })
+    const statementAtoms = [
+      <SubjectAtom
+        key={`subject-${statementIndex}`}
+        statements={statements}
+        choices={choices}
+        isFullWidth={isFullWidth}
+        statementIndex={statementIndex}
+        onChangeStatement={(value, structure) => {
+          this.handleChangeStatement(value, structure)
+          this.clearPredicate()
+        }}
+      />,
+      <VerbAtom
+        key={`verb-${statementIndex}`}
+        statements={statements}
+        choices={choices}
+        isFullWidth={isFullWidth}
+        statementIndex={statementIndex}
+        onChangeStatement={(value, structure) => {
+          this.handleChangeStatement(value, structure)
+        }}
+      />,
+      <ObjectAtom
+        key={`object-${statementIndex}`}
+        statements={statements}
+        choices={choices}
+        isFullWidth={isFullWidth}
+        statementIndex={statementIndex}
+      />,
+    ]
 
     return (
       <div>
@@ -239,7 +107,7 @@ class Statement extends React.Component {
                   <IconClose size={25} />
                 </div>
               )}
-            {statementAtoms}
+            {isRtl ? [...statementAtoms].reverse() : statementAtoms}
             {canDelete &&
               !isFullWidth &&
               !isRtl && (
@@ -287,7 +155,6 @@ Statement.defaultProps = {
   canDelete: true,
   statements: [{ subject: '', verb: '', object: null }],
   isRtl: false,
-  order: 'SVO',
   isFullWidth: false,
   statementIndex: 0,
   labels: { delete: 'DELETE' },
@@ -315,7 +182,7 @@ Statement.propTypes = {
   onChangeStatement: PropTypes.func,
   /** Statement remove callback */
   onRemoveStatement: PropTypes.func,
-  /** In which row does this Statement belong to?  */
+  /** To which row does this Statement belong to?  */
   statementIndex: PropTypes.number,
   /** Labels for the controls and texts, default is english */
   labels: PropTypes.shape({
